backend/requirements.txt
pandas
langchain
langchain-core
langchain-anthropic
langchain-xai
python-dotenv
fastapi
numpy
pydantic
recordlinkage
uvicorn